Initial upload of my SuperPy final project
In this commit, I'm uploading the final version of my SuperPy project to GitHub. This is the first time I'm adding the project to version control. I've also set the repository to public so that others can access and review the code. The readme and technical report can also be found in this repository.